attendance status is on teacher student model and appears in attendance view
